+++ mail/proton-mail-common/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Cache key for attachments into (u64, String)
 - Split `ContextualLabel` in `CustomFolder`, `CustomLabel` and `SystemLabel`. 
 - `Sidebar::custom_folders` to return all custom_folders in a hierarchical way.
 - Grouped `total_conv` and `total_msg` from `ContextualLabel` as `total`.